+++ tutorbranding/plugin.py
@@ -1,21 +1,14 @@
 """
 Tutor plugin to brand your Open edX instance.
 """
-import sys
 from glob import glob
 import os
-import zipfile
-from pathlib import Path
-from urllib.error import HTTPError
 
 import importlib_resources
-import click
-import requests
 from tutormfe.hooks import MFE_APPS
 
-
 from tutor import config as tutor_config
-from tutor import hooks, fmt
+from tutor import hooks
 
 from .__about__ import __version__
 
@@ -56,6 +49,10 @@ config = {
         # OVERRIDES: additional CSS for mfe branding
         "OVERRIDES": "",
         "FONTS": "",
+        "LMS_IMAGES": [],
+        "CMS_IMAGES": [],
+        "FONTS_URLS": [],
+
         "MFE": {},
         "MFE_LOGO_URL": '',
         "MFE_LOGO_WHITE_URL": '',
@@ -140,124 +137,6 @@ def _add_my_mfe(mfes):
 
     return mfes
 
-
-# Commands
-@click.group(help="Branding tools", name='branding')
-@click.pass_obj
-def branding_command(context):  # pylint: disable=unused-argument
-    """
-    Dummy function to group the branding commands
-    :return:
-    """
-
-
-def _download_file(url: str, dest_dir: str, filename: str = None) -> str:
-    """
-    Download a file from internet and store it in the destination directory.
-    :param url: where to get the file from
-    :param dest_dir: where to put the file
-    :return: file name.
-    """
-    fmt.echo_info(f"Downloading {url} to {dest_dir}")
-
-    try:
-        response = requests.get(url, timeout=10)
-        response.raise_for_status()
-
-        if not filename:
-            filename = os.path.basename(url)
-
-        with open(os.path.join(dest_dir, filename), "wb") as f:
-            f.write(response.content)
-
-    except HTTPError as http_err:
-        fmt.echo_error(f'HTTP error occurred downloading {url}: {http_err}')
-
-    return filename
-
-@click.command(help="Download image from url")
-@click.pass_obj
-def download_images(context):
-    """
-    Command to download images from a URL.
-    :param context: Click context
-    :return:
-    """
-    fmt.echo_info("*** Downloading images ***")
-    full_config = tutor_config.load(context.root)
-
-    # Download LMS images
-    dest_dir = os.path.join(context.root, 'env', 'build', 'openedx', 'themes',
-                            'theme', 'lms', 'static', 'images')
-
-    if "BRANDING_LMS_IMAGES" in full_config:
-        for image in full_config['BRANDING_LMS_IMAGES']:
-            Path(dest_dir).mkdir(parents=True, exist_ok=True)
-            _download_file(url=image.get('url'), filename=image.get('filename'), dest_dir=dest_dir)
-    else:
-        fmt.echo_alert("No BRANDING_LMS_IMAGES configured")
-
-    # Download CMS images
-    dest_dir = os.path.join(context.root, 'env', 'build', 'openedx', 'themes',
-                            'theme', 'cms', 'static', 'images')
-
-    if "BRANDING_CMS_IMAGES" in full_config:
-        Path(dest_dir).mkdir(parents=True, exist_ok=True)
-        for image in full_config['BRANDING_CMS_IMAGES']:
-            _download_file(url=image.get('url'), filename=image.get('filename'), dest_dir=dest_dir)
-    else:
-        fmt.echo_alert("No BRANDING_CMS_IMAGES configured")
-
-
-@click.command(help="Download and unzip font from url")
-@click.pass_obj
-def download_fonts(context):
-    """
-    Command to download fonts from a URL.
-    :param context: Click context
-    :return:
-    """
-    full_config = tutor_config.load(context.root)
-
-    # Download fonts
-    dest_dirs = [
-        os.path.join(context.root, 'env', 'build', 'openedx', 'themes',
-                     'theme', 'lms', 'static', 'fonts'),
-        os.path.join(context.root, 'env', 'plugins', 'mfe', 'build', 'mfe',
-                     'brand-openedx', 'fonts'),
-        os.path.join(context.root, 'env', 'plugins', 'mfe', 'build', 'mfe',
-                     'brand-openedx-learner-dashboard', 'fonts'),
-    ]
-
-    if "BRANDING_FONTS_URLS" in full_config:
-        for font_url in full_config['BRANDING_FONTS_URLS']:
-            for dest_dir in dest_dirs:
-                Path(dest_dir).mkdir(parents=True, exist_ok=True)
-                fmt.echo_info(f"Downloading fonts from {font_url}")
-                filename = _download_file(url=font_url, dest_dir=dest_dir)
-
-                if filename.endswith('.zip'):
-                    # Unzip the file
-                    fmt.echo_info(f"Unzipping font file {filename}")
-                    zipped_filename = os.path.join(dest_dir, filename)
-                    try:
-                        with zipfile.ZipFile(zipped_filename, 'r') as zipped_file:
-                            zipped_file.extractall(dest_dir)
-                            zipped_file.printdir()
-
-                            os.remove(os.path.join(dest_dir, filename))
-                    except zipfile.BadZipFile:
-                        fmt.echo_error(f"File {zipped_filename} is not a zip file.")
-                        sys.exit(1)
-
-    else:
-        fmt.echo_alert("No BRANDING_FONTS_URLS configured")
-
-
-branding_command.add_command(download_images)
-branding_command.add_command(download_fonts)
-
-hooks.Filters.CLI_COMMANDS.add_item(branding_command)
 
 # Load patches from files
 for path in glob(str(importlib_resources.files("tutorbranding") / "patches" / "*")):